requirements.txt
pytest==7.4.3
boto3==1.34.2
python-dotenv==1.0.0
logzero==1.7.0
jsonschema==4.20.0
pytest-xdist==3.5.0
faker==21.0.0
pyodbc==5.0.1
hypothesis==6.92.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
playwright==1.40.0
pytest-playwright==0.4.3